try to integrated preprocess into DataLoader and LazyDataLoader. Add some docstrings.
