+++ networks/layers/ssim.py
@@ -3,8 +3,6 @@ import warnings
 import torch
 from torch import nn
 
-from utils import data_related as dr
-
 
 def SSIM(y_hat: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
     # TODO：出现负数？
@@ -26,25 +24,6 @@ class SSIMLoss(nn.Module):
             if ssim < 0:
                 warnings.warn(f'出现了负值SSIM={ssim}！')
         return torch.mean(1 - ssim_of_each_sample)
-
-
-class Val2Fig(nn.Module):
-
-    def __init__(self, img_mode='L'):
-        """
-        将数值转化为要求模式的图片。
-        :param img_mode: 生成的图片模式。'L'为灰度图。
-        """
-        self.mode = img_mode
-        super().__init__()
-
-    def forward(self, y_hat: torch.Tensor):
-        # 进行归一化
-        y_hat = dr.normalize(y_hat)
-        if self.mode == 'L' or self.mode == 'RGB':
-            return (y_hat + 1) * 128
-        else:
-            return y_hat
