+++ utils/dataloader.py
@@ -28,8 +28,6 @@ class LazyDataLoader:
         self.__len = len(index_dataset) // batch_size
         self.__kwargs = kwargs
 
-        # self.__features_preprocesses = [] if features_preprocesses is None else features_preprocesses
-        # self.__labels_preprocess = [] if labels_preprocesses is None else labels_preprocesses
         self.__features_preprocesses = []
         self.__labels_preprocess = []
         # self.__index_loader = to_loader(index_dataset, batch_size * load_multiple, shuffle=shuffle)
@@ -38,17 +36,12 @@ class LazyDataLoader:
 
     def __iter__(self):
         for index, label in self.__index_loader:
+            # TODO：试图用多线程改进效率
             raw_ds = DataSet(self.__read_fn(index), label)
             # 进行预处理
             raw_ds.apply(self.__features_preprocesses, self.__labels_preprocess)
-            # batch_loader = raw_ds.to_loader(
-            #     self.__batch_size, self.__sampler, self.__shuffle, **self.__kwargs
-            # )
             # 此处不可以使用用户提供的sampler，会发生越界问题。此处数据集已经进行了打乱，因此shuffle操作也是不必要的
             batch_loader = raw_ds.to_loader(self.__batch_size, shuffle=False, **self.__kwargs)
-            # batch_loader = to_loader(
-            #     raw_ds, self.__batch_size, self.__sampler, self.__shuffle, **self.__kwargs
-            # )
             for X, y in batch_loader:
                 yield X, y
